// apps/admin/src/generated/api/schemas/uploadSettingsResponseDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { UploadSettingsResponseDtoStorageDriver } from "./uploadSettingsResponseDtoStorageDriver";

export interface UploadSettingsResponseDto {
  maxSizeMb: number;
  allowedMimeTypes: string[];
  environmentMaxSizeMb: number;
  environmentAllowedMimeTypes: string[];
  storageDriver: UploadSettingsResponseDtoStorageDriver;
}